requirements.txt
pandas>=1.3.0
numpy>=1.20.0
streamlit>=1.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
openai>=1.0.0
python-dotenv>=0.19.0
urllib3==1.26.18
six>=1.16.0
requests<3,>=2.27.0
sentence-transformers>=2.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
